# Optimize bottlenecks in 2.0 code