<!--
  Thank you for contributing! Please use this pull request (PR) template.


 In the description field of this PR, include "resolves #XXXX" tagging the issue you are fixing. If this PR addresses the issue but doesn't completely resolve it (ie the issue should remain open after your PR is merged), write "addresses #XXXX".-->
Resolves https://github.com/processing/p5.js/issues/8017
Resolves https://github.com/processing/p5.js/issues/8013

This implements some of the changes discussed in the above two issues.

Previously, this was the performance profile of time spent in functions:
<img width="430" height="251" alt="image" src="https://github.com/user-attachments/assets/01d8177c-7051-47c1-9719-c4e96ca90d2f" />

After these changes, it looks like this:
<img width="434" height="240" alt="image" src="https://github.com/user-attachments/assets/1f9c7e3a-9448-49a8-bf9a-1775c2f6dc0c" />

According to those, the largest problems were:
- Matrix constructors
- Vector constructors
- getters (I saw two: primarily the global binding getter, and to a lesser extent, one in p5.Vector.copy)

[On this test sketch](https://editor.p5js.org/davepagurek/sketches/uvm_o0k7-), starting with a base runtime of 420ms, here are the optimizations I made and the new runtimes afterwards:
- Remove the extra array copy from p5.Matrix constructor: 400ms
- Removes the Matrix interface method checks from the constructor: 390ms
- Caching bound functions on `p5.prototype`and constants that we make global: 315ms
- Removal of extra `get()`s in `p5.Vector` (using `.values` when `._values` is the same): 299ms
- Removal of extra `.map()`s in `p5.Vector` constructor: 292ms
- Removal of dynamic `typeof`s + caching bound functions in global non-`p5.prototype` functions: 250ms

For comparison, when using 1.11.0, the performance is 215ms. So this is still not *quite* at 1.x perf, but it's a lot closer.

## Takeaways

Dynamic getters are generally slower than property accesses! We should consider whether they're really necessary when we use them if it's something that will be accessed a lot, e.g. a constant.

Not from these particular tests, but at work we experimented in the past with using a `Proxy` to wrap p5 in order to get multiple global mode sketches to be able to run in a shared environment without them knowing, but found base performance proxying all p5 methods to be significantly worse. Rewriting the sketch code to use direct accesses gave comparable performance to regular global mode, despite the proxy methods all being pretty simple. So Proxies aren't an across-the-board replacement for dynamic getters either without measuring. In general we should just try to make as much as possible non-dynamic when it isn't necessary.