@davepagurek 2025-09-19
@limzykenneth could this possibly be an ordering issue? FES currently does the function validation wrapping in presetup, could we move the global binding to presetup too so that the order is like:
- FES presetup
- everything else?
- global binding comes last
...since addons presumably shouldn't be directly using globals anyway, and user sketch code should only be happening in setup onwards.